medapp_backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.2,<5.3
psycopg2-binary>=2.9,<3.0
boto3>=1.28,<1.29
requests>=2.28,<3.0
